Move oauth icons into central location
